# src/util/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Milk Hiring",
        "version": "1.0.0",
        "description": "API para gestão de fazendas e produção de leite"
    },
...
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "Servidor local"
        }
    ],
    "components": {
        "schemas": {
            "Farmer": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "671d4f77e17ad949a20ffbb0"
                    },
                    "name": {
                        "type": "string",
                        "example": "João"
                    },
                    "birthDate": {
                        "type": "string",
                        "example": "1990-01-01"
                    },
                    "address": {
                        "type": "object",
                        "properties": {
                            "street": {
                                "type": "string",
                                "example": "Rua 1"
                            },
                            "number": {
                                "type": "string",
                                "example": "123"
                            },
                            "city": {
                                "type": "string",
                                "example": "Cidade 1"
                            },
                            "state": {
                                "type": "string",
                                "example": "Estado 1"
                            },
                            "neighborhood": {
                                "type": "string",
                                "example": "Bairro 1"
                            },
                            "cep": {
                                "type": "string",
                                "example": "12345678"
                            }
                        }
                    },
                    "documentNumber": {
                        "type": "string",
                        "example": "11122233344"
                    },
                    "phone": {
                        "type": "string",
                        "example": "11970707070"
                    },
                    "email": {
                        "type": "string",
                        "example": "joao@example.com"
                    }
                }
            },
...
            "Farm": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "671d4f77e17ad949a20ffbb0"
                    },
                    "name": {
                        "type": "string",
                        "example": "Fazenda 1"
                    },
                    "address": {
                        "type": "object",
                        "properties": {
                            "street": {
                                "type": "string",
                                "example": "Rua 1"
                            },
                            "number": {
                                "type": "string",
                                "example": "123"
                            },
                            "city": {
                                "type": "string",
                                "example": "Cidade 1"
                            },
                            "state": {
                                "type": "string",
                                "example": "Estado 1"
                            },
                            "neighborhood": {
                                "type": "string",
                                "example": "Bairro 1"
                            },
                            "cep": {
                                "type": "string",
                                "example": "12345678"
                            }
                        }
                    },
                    "distanceFarmToFactory": {
                        "type": "number",
                        "example": 10
                    },
                    "farmer": {
                        "$ref": "#/components/schemas/Farmer"
                    }
                }
            },
...
            "MilkProduction": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "671d4f77e17ad949a20ffbb0"
                    },
                    "date": {
                        "type": "string",
                        "example": "2022-01-01"
                    },
                    "quantity": {
                        "type": "number",
                        "example": 10
                    },
                    "farm": {
                        "$ref": "#/components/schemas/Farm"
                    }
                }
            },
            "MilkProductionInput": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string",
                        "example": "2022-01-01T00:00:00.000Z"
                    },
                    "quantity": {
                        "type": "number",
                        "example": 10
                    },
                    "farmId": {
                        "type": "string",
                        "example": "671d4f77e17ad949a20ffbb0"
                    }
                },
                "required": [
                    "date",
                    "quantity",
                    "farmId"
                ]
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Erro"
                    }
                }
            }
        }
    },
    "paths": {
...
        "/milk-production/{id}": {
            "get": {
                "tags": [
                    "Milk Production Management"
                ],
                "summary": "Obter Produção de Leite por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produção de leite encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MilkProduction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fazenda não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Milk Production Management"
                ],
                "summary": "Atualizar Fazenda por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MilkProductionInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fazenda atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MilkProduction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Fazenda não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Milk Production Management"
                ],
                "summary": "Deletar Produção de Leite por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produção de Leite deletada com sucesso"
                    },
                    "404": {
                        "description": "Produção de Leite não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}